% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop.R
\name{pop}
\alias{pop}
\title{Pop the first element in-place}
\usage{
pop(x)
}
\arguments{
\item{x}{'stack' from which a value should be popped}
}
\value{
the popped first element of \code{x}
}
\description{
Pop the first element in-place
}
\details{
This dispatches based on the \code{class()} of \code{x}. A macro
is then used to redefine \code{x} in the \code{parent.frame()} such that the
first element (as defined in that \code{class}) is removed in-place. This first
element is returned.
}
