% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{push}
\alias{push}
\title{Push an element to an object in-place}
\usage{
push(x, y, ...)
}
\arguments{
\item{x}{'stack' into which a value should be pushed}

\item{y}{value (matching the class of \code{x}) to be pushed on to \code{x}}

\item{...}{unused}
}
\value{
the updated object \code{x}, invisibly
}
\description{
Push an element to an object in-place
}
\details{
This dispatches based on the \code{class()} of \code{x}. A macro
is then used to redefine \code{x} in the \code{parent.frame()} such that \code{y} is inserted
as the first element in-place.
}
