% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-utils.R, R/02-branding.R, R/05-scope.R
\name{weasel-deprecated}
\alias{weasel-deprecated}
\alias{logo}
\alias{reshape_to_wide}
\alias{summarize_waves}
\alias{filter_wave_summary}
\alias{get_data_by_row}
\alias{generate_sets}
\alias{filter_sets}
\title{Deprecated functions}
\usage{
logo(...)

reshape_to_wide(...)

summarize_waves(...)

filter_wave_summary(...)

get_data_by_row(...)

generate_sets()

filter_sets()
}
\arguments{
\item{...}{Arguments passed on to the replacement function.}
}
\value{
The value of the corresponding replacement function.
}
\description{
These earlier names are kept so existing scripts keep running, but they
now emit a deprecation warning (class \code{weasel_deprecated}) and forward
to their replacements. They will be removed in a future release.
}
\details{
\itemize{
\item \code{reshape_to_wide()} -> \code{\link[=weasel_reshape_to_wide]{weasel_reshape_to_wide()}}
\item \code{summarize_waves()} -> \code{\link[=weasel_summarize_waves]{weasel_summarize_waves()}}
\item \code{filter_wave_summary()} -> \code{\link[=weasel_filter_wave_summary]{weasel_filter_wave_summary()}}
\item \code{get_data_by_row()} -> \code{\link[=weasel_get_data_by_row]{weasel_get_data_by_row()}}
\item \code{logo()} -> \code{\link[=weasel_logo]{weasel_logo()}}
\item \code{generate_sets()} and \code{filter_sets()} remain no-ops (see
\code{\link[=evaluate_weasel_scope]{evaluate_weasel_scope()}}) and now warn as well.
}
}
\keyword{internal}
