% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weasel-package.R
\docType{package}
\name{weasel-package}
\alias{weasel}
\alias{weasel-package}
\title{weasel: Provides `pop()` and `push()` functionality}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides `pop()` and `push()` functionality via macros which remove or add the first element in-place.
}
\author{
\strong{Maintainer}: Jonathan Carroll \email{rpkg@jcarroll.com.au} (\href{https://orcid.org/0000-0002-1404-5264}{ORCID})

}
\keyword{internal}
